cv-final/cv-final/requirements.txt
flask==3.0.0
tensorflow==2.15.0
numpy>=1.24.0
opencv-python>=4.8.1
pillow>=10.0.0